backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
email-validator==2.1.0
aiosmtplib==3.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.109.0
dlib==19.24.2
face-recognition==1.3.0
imutils==0.5.4
